pyproject.toml
[project]
name = "magicstory-cli"
version = "0.1.0"
description = "CLI pipeline for generating illustrated storybooks and PDFs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "pydantic>=2.9.0",
  "pyyaml>=6.0.2",
  "typer>=0.12.5",
  "rich>=13.9.4",
  "weasyprint>=64.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.6.9",
]

[project.scripts]
story = "magicstory_cli.cli.app:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/magicstory_cli"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
